Move run_custom_user_commands to fix an import cycle
